Move some code out of state.go, no functional changes
