+++ state.go
@@ -9,24 +9,15 @@ import (
 	"sync/atomic"
 	"time"
 
-	"github.com/hashicorp/go-multierror"
+	filaddr "github.com/filecoin-project/go-address"
+	"github.com/filecoin-project/go-state-types/abi"
 	"github.com/filecoin-project/go-state-types/big"
+	filbig "github.com/filecoin-project/go-state-types/big"
 	lotusbs "github.com/filecoin-project/lotus/blockstore"
 	"github.com/filecoin-project/lotus/build"
 	"github.com/filecoin-project/lotus/chain/actors"
 	"github.com/filecoin-project/lotus/chain/types/ethtypes"
-	"github.com/hashicorp/go-multierror"
-	leveldb "github.com/ipfs/go-ds-leveldb"
-	exchange "github.com/ipfs/go-ipfs-exchange-interface"
-	ipld "github.com/ipfs/go-ipld-format"
-	"github.com/ipfs/go-verifcid"
 	carbs "github.com/ipld/go-car/v2/blockstore"
-	"github.com/libp2p/go-libp2p"
-	"github.com/libp2p/go-libp2p/core/network"
-
-	filaddr "github.com/filecoin-project/go-address"
-	"github.com/filecoin-project/go-state-types/abi"
-	filbig "github.com/filecoin-project/go-state-types/big"
 
 	lchadt "github.com/filecoin-project/lotus/chain/actors/adt"
 	lbi "github.com/filecoin-project/lotus/chain/actors/builtin"
@@ -39,7 +30,6 @@ import (
 	"github.com/filecoin-project/go-state-types/builtin/v10/util/adt"
 	cbg "github.com/whyrusleeping/cbor-gen"
 
-	blocks "github.com/ipfs/go-block-format"
 	"github.com/ipfs/go-cid"
 	blockstore "github.com/ipfs/go-ipfs-blockstore"
 	ipldcbor "github.com/ipfs/go-ipld-cbor"
@@ -47,241 +37,6 @@ import (
 	"golang.org/x/sync/errgroup"
 	"golang.org/x/xerrors"
 )
-
-type countingBlockstore struct {
-	blockstore.Blockstore
-	mx          sync.Mutex
-	m           map[cid.Cid]int
-	orderedCids []cid.Cid
-}
-
-func (c *countingBlockstore) Get(ctx context.Context, cid cid.Cid) (blocks.Block, error) {
-	blk, err := c.Blockstore.Get(ctx, cid)
-	if err != nil {
-		return nil, err
-	}
-
-	c.mx.Lock()
-	_, found := c.m[cid]
-	c.m[cid] = len(blk.RawData())
-	if !found {
-		c.orderedCids = append(c.orderedCids, cid)
-	}
-	c.mx.Unlock()
-
-	return blk, nil
-}
-
-var _ blockstore.Blockstore = (*countingBlockstore)(nil)
-
-func getStateFromCar(ctx context.Context, srcSnapshot string) (blockstore.Blockstore, lchtypes.TipSetKey, error) {
-	start := time.Now()
-	defer func() {
-		fmt.Printf("duration: %v\n", time.Since(start))
-	}()
-
-	carbs, err := blockstoreFromSnapshot(ctx, srcSnapshot)
-	if err != nil {
-		return nil, lchtypes.EmptyTSK, err
-	}
-
-	fmt.Printf("duration to load snapshot: %v\n", time.Since(start))
-
-	var tsk lchtypes.TipSetKey
-	carRoots, err := carbs.Roots()
-	if err != nil {
-		return nil, lchtypes.EmptyTSK, err
-	}
-	tsk = lchtypes.NewTipSetKey(carRoots...)
-
-	return carbs, tsk, nil
-}
-
-func getStateDynamicallyLoadedFromBitswap(ctx context.Context) (blockstore.Blockstore, error) {
-	start := time.Now()
-	defer func() {
-		fmt.Printf("duration: %v\n", time.Since(start))
-	}()
-
-	h, err := libp2p.New(libp2p.ResourceManager(&network.NullResourceManager{}))
-	if err != nil {
-		return nil, err
-	}
-
-	bscFil, err := setupFilContentFetching(h, ctx)
-	if err != nil {
-		return nil, err
-	}
-	bscIpfs, err := setupIPFSContentFetching(h, ctx)
-	if err != nil {
-		return nil, err
-	}
-
-	lds, err := leveldb.NewDatastore("", nil)
-	if err != nil {
-		return nil, err
-	}
-	baseBstore := blockstore.NewBlockstore(lds)
-	cf := &combineFetcher{
-		0, 0,
-		bscFil.NewSession(ctx),
-		bscIpfs.NewSession(ctx),
-	}
-	bs := blockstore.NewIdStore(&bservBstoreWrapper{Blockstore: baseBstore, bserv: cf})
-
-	go func() {
-		ticker := time.NewTicker(10 * time.Second)
-		defer ticker.Stop()
-		for {
-			select {
-			case <-ctx.Done():
-				return
-			case <-ticker.C:
-				peers := h.Network().Peers()
-				nPeersWithFilBitswap := 0
-				nPeersWithIpfsBitswap := 0
-				for _, p := range peers {
-					const ipfsBs = "/ipfs/bitswap/1.2.0"
-					const filBs = "/chain" + ipfsBs
-					retProto, err := h.Peerstore().FirstSupportedProtocol(p, filBs, ipfsBs)
-					if err != nil || retProto == "" {
-						continue
-					}
-
-					//TODO: This isn't technically correct since peers can support both, but in practice none do
-					switch retProto {
-					case filBs:
-						nPeersWithFilBitswap++
-					case ipfsBs:
-						nPeersWithIpfsBitswap++
-					}
-				}
-				nwantsFil := len(bscFil.GetWantlist())
-				nwantsIpfs := len(bscIpfs.GetWantlist())
-				nFilBlks := atomic.LoadUint64(&cf.numFirstBlocks)
-				nIpfsBlks := atomic.LoadUint64(&cf.numSecondBlocks)
-				fmt.Printf("numPeers: %d, "+
-					"nPeersWithFilBitswap: %d, numWantsFil: %d, nblksFil: %d, "+
-					"nPeersWithIpfsBitswap: %d, numWantsIpfs: %d, nblksIpfs: %d \n",
-					len(peers), nPeersWithFilBitswap, nwantsFil, nFilBlks, nPeersWithIpfsBitswap, nwantsIpfs, nIpfsBlks)
-			}
-		}
-	}()
-
-	fmt.Printf("duration to setup bitswap fetching: %v\n", time.Since(start))
-
-	return bs, nil
-}
-
-type combineFetcher struct {
-	numFirstBlocks, numSecondBlocks uint64
-	first, second                   exchange.Fetcher
-}
-
-func (f *combineFetcher) GetBlock(ctx context.Context, c cid.Cid) (blocks.Block, error) {
-	subCtx, cancel := context.WithCancel(ctx)
-	defer cancel()
-	firstCh, firstErr := f.first.GetBlocks(subCtx, []cid.Cid{c})
-	secondCh, secondErr := f.second.GetBlocks(subCtx, []cid.Cid{c})
-	if firstErr != nil && secondErr != nil {
-		return nil, multierror.Append(firstErr, secondErr)
-	}
-	if firstErr != nil {
-		blk := <-secondCh
-		atomic.AddUint64(&f.numSecondBlocks, 1)
-		return blk, nil
-	}
-	if secondErr != nil {
-		blk := <-firstCh
-		atomic.AddUint64(&f.numFirstBlocks, 1)
-		return blk, nil
-	}
-
-	var blk blocks.Block
-	select {
-	case blk = <-firstCh:
-		atomic.AddUint64(&f.numFirstBlocks, 1)
-	case blk = <-secondCh:
-		atomic.AddUint64(&f.numSecondBlocks, 1)
-	case <-ctx.Done():
-		return nil, ctx.Err()
-	}
-	if blk == nil {
-		if ctx.Err() != nil {
-			return nil, ctx.Err()
-		}
-		return nil, fmt.Errorf("block missing... this shouldn't happen. Report an error")
-	}
-	return blk, nil
-}
-
-func (f *combineFetcher) GetBlocks(ctx context.Context, cids []cid.Cid) (<-chan blocks.Block, error) {
-	//TODO implement me
-	panic("implement me")
-}
-
-var _ exchange.Fetcher = (*combineFetcher)(nil)
-
-// this is the wrong way around, but this is the easiest way to hack it
-type bservBstoreWrapper struct {
-	blockstore.Blockstore
-	bserv exchange.Fetcher
-}
-
-func (b *bservBstoreWrapper) Get(ctx context.Context, c cid.Cid) (blocks.Block, error) {
-	err := verifcid.ValidateCid(c) // hash security
-	if err != nil {
-		return nil, err
-	}
-
-	block, err := b.Blockstore.Get(ctx, c)
-	if err == nil {
-		return block, nil
-	}
-
-	if ipld.IsNotFound(err) {
-		// TODO be careful checking ErrNotFound. If the underlying
-		// implementation changes, this will break.
-		blk, err := b.bserv.GetBlock(ctx, c)
-		if err != nil {
-			return nil, err
-		}
-		// also write in the blockstore for caching
-		err = b.Blockstore.Put(ctx, blk)
-		if err != nil {
-			return nil, err
-		}
-		return blk, nil
-	}
-
-	return nil, err
-}
-
-func (b *bservBstoreWrapper) GetSize(ctx context.Context, c cid.Cid) (int, error) {
-	has, err := b.Blockstore.Has(ctx, c)
-	if has {
-		return b.Blockstore.GetSize(ctx, c)
-	}
-	if !ipld.IsNotFound(err) {
-		// TODO be careful checking ErrNotFound. If the underlying
-		// implementation changes, this will break.
-		return 0, err
-	}
-
-	blk, err := b.bserv.GetBlock(ctx, c)
-	if err != nil {
-		return 0, err
-	}
-	// also write in the blockstore for caching
-	err = b.Blockstore.Put(ctx, blk)
-	if err != nil {
-		return 0, err
-	}
-
-	return len(blk.RawData()), nil
-}
-
-var _ blockstore.Blockstore = (*bservBstoreWrapper)(nil)
 
 func getCoins(ctx context.Context, bstore blockstore.Blockstore, tsk lchtypes.TipSetKey, addr filaddr.Address) (defErr error) {
 	cbs := &countingBlockstore{Blockstore: bstore, m: make(map[cid.Cid]int)}
@@ -331,51 +86,6 @@ func getCoins(ctx context.Context, bstore blockstore.Blockstore, tsk lchtypes.Ti
 	fmt.Printf("total block sizes: %d\n", totalSizeBytes)
 	return
 }
-
-type getManyCborStore struct {
-	*ipldcbor.BasicIpldStore
-}
-
-func (g *getManyCborStore) GetMany(ctx context.Context, cids []cid.Cid, outs []interface{}) <-chan *hamt.OptionalInteger {
-	outCh := make(chan *hamt.OptionalInteger)
-	wg := sync.WaitGroup{}
-	processingCh := make(chan int)
-	go func() {
-		for i := 0; i < len(cids); i++ {
-			processingCh <- i
-		}
-		close(processingCh)
-	}()
-	const concurrency = 8
-	wg.Add(concurrency)
-	for i := 0; i < concurrency; i++ {
-		go func() {
-			defer wg.Done()
-			for index := range processingCh {
-				c := cids[index]
-				o := outs[index]
-				err := g.Get(ctx, c, o)
-				if err != nil {
-					outCh <- &hamt.OptionalInteger{Error: err}
-				} else {
-					outCh <- &hamt.OptionalInteger{Value: index}
-				}
-			}
-		}()
-	}
-
-	go func() {
-		wg.Wait()
-		close(outCh)
-	}()
-	return outCh
-}
-
-type getManyIPLDStore interface {
-	GetMany(ctx context.Context, cids []cid.Cid, outs []interface{}) <-chan *hamt.OptionalInteger
-}
-
-var _ getManyIPLDStore = (*getManyCborStore)(nil)
 
 func parseActors(ctx context.Context, sm *lchstmgr.StateManager, ts *lchtypes.TipSet, rootAddr filaddr.Address, foundAttoFil abi.TokenAmount) error {
 	ast := lchadt.WrapStore(ctx, ipldcbor.NewCborStore(sm.ChainStore().UnionStore()))
