+++ util.go
@@ -2,23 +2,30 @@ package main
 
 import (
 	"context"
+	"fmt"
+	"log"
+	"os"
+	"sync"
+	"time"
+
 	"github.com/filecoin-project/go-address"
 	lotusbs "github.com/filecoin-project/lotus/blockstore"
 	"github.com/filecoin-project/lotus/chain/consensus"
 	"github.com/filecoin-project/lotus/chain/consensus/filcns"
 	"github.com/filecoin-project/lotus/chain/stmgr"
 	chainstore "github.com/filecoin-project/lotus/chain/store"
+	lchtypes "github.com/filecoin-project/lotus/chain/types"
 	"github.com/filecoin-project/lotus/chain/vm"
 	"github.com/filecoin-project/lotus/cmd/lotus-sim/simulation/mock"
+	blocks "github.com/ipfs/go-block-format"
 	"github.com/ipfs/go-cid"
 	ds "github.com/ipfs/go-datastore"
-	"github.com/ipfs/go-datastore/sync"
+	dssync "github.com/ipfs/go-datastore/sync"
+	blockstore "github.com/ipfs/go-ipfs-blockstore"
 	"github.com/ipld/go-car/v2"
 	carbs "github.com/ipld/go-car/v2/blockstore"
 	caridx "github.com/ipld/go-car/v2/index"
 	"golang.org/x/xerrors"
-	"log"
-	"os"
 )
 
 func cidMustParse(s string) cid.Cid {
@@ -38,7 +45,7 @@ func mustAddrID(a address.Address) uint64 {
 }
 
 func newFilStateReader(bs lotusbs.Blockstore) (*stmgr.StateManager, error) {
-	mds := sync.MutexWrap(ds.NewMapDatastore())
+	mds := dssync.MutexWrap(ds.NewMapDatastore())
 	c := cid.MustParse("bafy2bzacecnamqgqmifpluoeldx7zzglxcljo6oja4vrmtj7432rphldpdmm2")
 	err := mds.Put(context.TODO(), ds.NewKey("0"), c.Bytes())
 	if err != nil {
@@ -64,7 +71,30 @@ func newFilStateReader(bs lotusbs.Blockstore) (*stmgr.StateManager, error) {
 	)
 }
 
-func blockstoreFromSnapshot(ctx context.Context, snapshotFilename string) (*carbs.ReadOnly, error) {
+func getStateFromCar(ctx context.Context, srcSnapshot string) (blockstore.Blockstore, lchtypes.TipSetKey, error) {
+	start := time.Now()
+	defer func() {
+		fmt.Printf("duration: %v\n", time.Since(start))
+	}()
+
+	carbs, err := blockstoreFromSnapshot(ctx, srcSnapshot)
+	if err != nil {
+		return nil, lchtypes.EmptyTSK, err
+	}
+
+	fmt.Printf("duration to load snapshot: %v\n", time.Since(start))
+
+	var tsk lchtypes.TipSetKey
+	carRoots, err := carbs.Roots()
+	if err != nil {
+		return nil, lchtypes.EmptyTSK, err
+	}
+	tsk = lchtypes.NewTipSetKey(carRoots...)
+
+	return carbs, tsk, nil
+}
+
+func blockstoreFromSnapshot(_ context.Context, snapshotFilename string) (*carbs.ReadOnly, error) {
 	carFile := snapshotFilename
 	carFh, err := os.Open(carFile)
 	if err != nil {
@@ -106,3 +136,29 @@ func blockstoreFromSnapshot(ctx context.Context, snapshotFilename string) (*carb
 
 	return roBs, nil
 }
+
+type countingBlockstore struct {
+	blockstore.Blockstore
+	mx          sync.Mutex
+	m           map[cid.Cid]int
+	orderedCids []cid.Cid
+}
+
+func (c *countingBlockstore) Get(ctx context.Context, cid cid.Cid) (blocks.Block, error) {
+	blk, err := c.Blockstore.Get(ctx, cid)
+	if err != nil {
+		return nil, err
+	}
+
+	c.mx.Lock()
+	_, found := c.m[cid]
+	c.m[cid] = len(blk.RawData())
+	if !found {
+		c.orderedCids = append(c.orderedCids, cid)
+	}
+	c.mx.Unlock()
+
+	return blk, nil
+}
+
+var _ blockstore.Blockstore = (*countingBlockstore)(nil)
